enabled communication with shoppingcart stateful service, and added some of database models

+++ OnlineStore/OnlineStore.Communication/UserManagement/Models/UserManagementUserGetByIdResponseModel.cs
@@ -12,7 +12,7 @@
 
         public DateTime? DateOfBirth { get; set; }
 
-        public int? Age { get; set; }
+        public long? Age { get; set; }
 
         public string? Address { get; set; }
